refactor: removed sections.tsx
