--- models/sections.tsx
@@ -1,28 +1,0 @@
-import React, { ReactElement } from 'react';
-import LandingPageHeaders from '../components/landingPage/LandingPageHeaders'
-import AboutMe from '../components/about/AboutMe'
-import SocialsTray from '../components/about/SocialsTray'
-import SkillsTray from '../components/about/SkillsTray'
-import ProjectsTray from '../components/projects/ProjectsTray'
-import CountryCardTray from 'components/travel/CountryCardTray'
-import ContactSection from 'components/contact/ContactSection'
-
-export interface Section {
-  margin?: string;
-  sectionTitle?: string;
-  sectionTitleMargin?: string;
-  coreComponent: ReactElement;
-  trayComponent?: ReactElement;
-  pageNumber?: string;
-  nextPageReference?: string;
-  nextButtonMargins?: string;
-}
-
-export const sections: Map<string, Section> = new Map([
-    ["home", {coreComponent: <LandingPageHeaders/>, trayComponent: <SocialsTray/>, nextPageReference: "#about", nextButtonMargins: "mt-[19rem] lg:mt-[15rem] xl:mt-[21.5rem] 2xl:mt-[21%]"}],
-    ["about", { sectionTitle: "About Me", coreComponent: <AboutMe/>, trayComponent: <SkillsTray/>, pageNumber: "١", nextPageReference: "#projects", nextButtonMargins: "mt-2.5 2xl:mt-[11.5%] 2xl:mb-36"}],
-    ["projects", {sectionTitle: "Projects", sectionTitleMargin: "mb-7", coreComponent: <ProjectsTray/>, pageNumber: "٢", nextPageReference: "#travel", nextButtonMargins: "mt-2.5 lg:mt-72 lg:mb-16 2xl:mt-[36%]"}],
-    ["travel", {margin: "mb-72 mt-[40rem]", sectionTitle: "Travel", sectionTitleMargin: "mb-7", coreComponent: <CountryCardTray/>, pageNumber: "٣", nextButtonMargins: "" ,nextPageReference: "#contact"}],
-    ["contact", {margin: "mb-72 ", sectionTitle: "Get in touch!", sectionTitleMargin: "mb-7", coreComponent: <ContactSection/>, pageNumber: "٤"}
-  ]
-  ]);
